Refactor policy HUD rendering
